requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-channels==4.0.0
Pillow==9.5.0
python-dotenv==1.0.0
daphne==4.0.0
pytest==7.4.3
pytest-django==4.5.2
gunicorn==21.2.0